update to verify if the simulation date range is included in or the same as the input period
Loop added to verify if the simulation date range is included in or the same as the input period given to the model. If the simulation range only includes part of the input range, the script will only consider those dates when formatting the input. If the simulation range includes dates outside of the input range, return an error.